requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachelib==0.9.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decouple==0.0.7
Flask==2.2.2
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Share==0.1.1
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
psycopg2==2.9.6
pypng==0.20220715.0
pyshortner==0.0.4
python-dotenv==0.21.1
qrcode==7.4.2
redis==4.5.5
shortuuid==1.0.11
shortzyy==0.0.6
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.2.3
yarl==1.9.2
